requirements.txt
flask>=3.0.0
flask-cors>=4.0.0
python-pptx>=0.6.21
pydantic>=2.0.0
langchain>=0.1.0
langchain-openai>=0.0.5
requests>=2.31.0
